max_stack: limit value sizes to u8
This gives enough space for 256-byte values, which is probably more than
plenty. In return we reduce the amount of data we store for tracking
operand sizes 8-fold. Cache lines are happy.

+++ src/test.rs
@@ -293,7 +293,7 @@ impl<'a> TestContext {
 
             struct DefaultConfig;
             impl AnalysisConfig for DefaultConfig {
-                fn size_of_value(&self, ty: wasmparser::ValType) -> u64 {
+                fn size_of_value(&self, ty: wasmparser::ValType) -> u8 {
                     use wasmparser::ValType::*;
                     match ty {
                         I32 => 4,
